Extract draft token constants to pixi-free config
tokens.test.ts pulled STARTING_DRAFT_TOKENS / REROLL_TOKEN_COST from
scenes/play.ts, which transitively imports pixi.js. In the CI node env
pixi's BrowserAdapter dereferences navigator at module load and the
test file fails to even parse. Move the two constants to game/config.ts
so tests stay pixi-free.
