+++ axiom/src/game/config.ts
@@ -31,3 +31,7 @@ export const MAX_FRAME = 0.1;
 
 // Hit flash (seconds)
 export const HIT_FLASH_TIME = 0.1;
+
+// Draft token economy (concept.md § Resource model).
+export const STARTING_DRAFT_TOKENS = 2;
+export const REROLL_TOKEN_COST = 1;
